requirements.txt
Django==2.2.4
djangorestframework==3.10.2
factory-boy==2.12.0
Faker==2.0.1
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
